+++ test/async_notifier_test.dart
@@ -11,8 +11,8 @@ void main() {
     int? data = 0;
     final n = AsyncNotifier<int>(data: data);
     expect(data, 0);
-    expect(n.data, 0);
-    expect(n, const AsyncSnapshot.withData(ConnectionState.none, 0));
+    expect(n.snapshot.data, 0);
+    expect(n.snapshot, const AsyncSnapshot.withData(ConnectionState.none, 0));
   });
 
   group('AsyncNotifier.future', () {
@@ -20,24 +20,24 @@ void main() {
     late final n = AsyncNotifier<int>()..listen((e) => data = e.data);
     test('ConnectionState.none', () async {
       expect(data, null);
-      expect(n.data, null);
-      expect(n.isLoading, false);
-      expect(n, const AsyncSnapshot<int>.nothing());
+      expect(n.snapshot.data, null);
+      expect(n.snapshot.isLoading, false);
+      expect(n.snapshot, const AsyncSnapshot<int>.nothing());
     });
 
     test('ConnectionState.waiting', () async {
       n.future = Future.delayed(const Duration(milliseconds: 100), () => 1);
       expect(data, null);
-      expect(n.data, null);
-      expect(n.isLoading, true);
-      expect(n, const AsyncSnapshot<int>.waiting());
+      expect(n.snapshot.data, null);
+      expect(n.snapshot.isLoading, true);
+      expect(n.snapshot, const AsyncSnapshot<int>.waiting());
     });
 
     test('ConnectionState.done', () async {
       await Future<void>.delayed(const Duration(milliseconds: 200));
       expect(data, 1);
-      expect(n.data, 1);
-      expect(n, AsyncSnapshot.withData(ConnectionState.done, data));
+      expect(n.snapshot.data, 1);
+      expect(n.snapshot, AsyncSnapshot.withData(ConnectionState.done, data));
     });
   });
 
@@ -58,28 +58,28 @@ void main() {
 
     n.future = Future.delayed(const Duration(seconds: 1), () => throw e);
     expect(data, null);
-    expect(n.data, null);
-    expect(n.isLoading, true);
-    expect(n.hasError, false);
-    expect(n.error, null);
-    expect(n.connectionState, ConnectionState.waiting);
-    expect(n, const AsyncSnapshot<bool>.waiting());
+    expect(n.snapshot.data, null);
+    expect(n.snapshot.isLoading, true);
+    expect(n.snapshot.hasError, false);
+    expect(n.snapshot.error, null);
+    expect(n.snapshot.connectionState, ConnectionState.waiting);
+    expect(n.snapshot, const AsyncSnapshot<bool>.waiting());
 
     await Future<void>.delayed(const Duration(seconds: 1));
     expect(data, null);
-    expect(n.data, null);
-    expect(n.isLoading, false);
-    expect(n.hasError, true);
-    expect(n.error, error);
-    expect(n.error, isA<Error>());
-    expect(n.connectionState, ConnectionState.done);
+    expect(n.snapshot.data, null);
+    expect(n.snapshot.isLoading, false);
+    expect(n.snapshot.hasError, true);
+    expect(n.snapshot.error, error);
+    expect(n.snapshot.error, isA<Error>());
+    expect(n.snapshot.connectionState, ConnectionState.done);
 
     final snapshotWithError = AsyncSnapshot<bool>.withError(
       ConnectionState.done,
       error!,
       stackTrace!,
     );
-    expect(n, snapshotWithError);
+    expect(n.snapshot, snapshotWithError);
   });
 
   test('AsyncNotifier.future with new future', () async {
@@ -89,30 +89,30 @@ void main() {
     // First future.
     n.future = Future.delayed(const Duration(milliseconds: 100), () => 1);
     expect(data, null);
-    expect(n.data, null);
-    expect(n.isLoading, true);
-    expect(n, const AsyncSnapshot<int>.waiting());
+    expect(n.snapshot.data, null);
+    expect(n.snapshot.isLoading, true);
+    expect(n.snapshot, const AsyncSnapshot<int>.waiting());
 
     // Second future.
     n.future = Future.delayed(const Duration(milliseconds: 200), () => 2);
     expect(data, null);
-    expect(n.data, null);
-    expect(n.isLoading, true);
-    expect(n, const AsyncSnapshot<int>.waiting());
+    expect(n.snapshot.data, null);
+    expect(n.snapshot.isLoading, true);
+    expect(n.snapshot, const AsyncSnapshot<int>.waiting());
 
     // First future has no time to complete. It's overriden by second future.
     await Future<void>.delayed(const Duration(milliseconds: 150));
     expect(data, isNot(1));
-    expect(n.data, isNot(1));
-    expect(n.isLoading, true);
-    expect(n, const AsyncSnapshot<int>.waiting());
+    expect(n.snapshot.data, isNot(1));
+    expect(n.snapshot.isLoading, true);
+    expect(n.snapshot, const AsyncSnapshot<int>.waiting());
 
     // Second future completes normally.
     await Future<void>.delayed(const Duration(milliseconds: 200));
     expect(data, 2);
-    expect(n.data, 2);
-    expect(n.isLoading, false);
-    expect(n, AsyncSnapshot.withData(ConnectionState.done, data));
+    expect(n.snapshot.data, 2);
+    expect(n.snapshot.isLoading, false);
+    expect(n.snapshot, AsyncSnapshot.withData(ConnectionState.done, data));
   });
 
   group('AsyncNotifier.stream', () {
@@ -120,32 +120,32 @@ void main() {
     final n = AsyncNotifier<int>()..listen((e) => data = e.data);
     test('ConnectionState.none', () async {
       expect(data, null);
-      expect(n.data, null);
-      expect(n.isLoading, false);
-      expect(n, const AsyncSnapshot<int>.nothing());
+      expect(n.snapshot.data, null);
+      expect(n.snapshot.isLoading, false);
+      expect(n.snapshot, const AsyncSnapshot<int>.nothing());
     });
 
     test('ConnectionState.waiting', () async {
       /// will stream 0, 1, 2
       n.stream = Stream.periodic(const Duration(seconds: 1), (i) => i).take(3);
       expect(data, null);
-      expect(n.data, null);
-      expect(n.isLoading, true);
-      expect(n, const AsyncSnapshot<int>.waiting());
+      expect(n.snapshot.data, null);
+      expect(n.snapshot.isLoading, true);
+      expect(n.snapshot, const AsyncSnapshot<int>.waiting());
     });
 
     test('ConnectionState.active', () async {
       await Future<void>.delayed(const Duration(seconds: 2));
       expect(data, 1);
-      expect(n.data, 1);
-      expect(n, AsyncSnapshot.withData(ConnectionState.active, data));
+      expect(n.snapshot.data, 1);
+      expect(n.snapshot, AsyncSnapshot.withData(ConnectionState.active, data));
     });
 
     test('ConnectionState.done', () async {
       await Future<void>.delayed(const Duration(seconds: 3));
       expect(data, 2);
-      expect(n.data, 2);
-      expect(n, AsyncSnapshot.withData(ConnectionState.done, data));
+      expect(n.snapshot.data, 2);
+      expect(n.snapshot, AsyncSnapshot.withData(ConnectionState.done, data));
     });
   });
 
@@ -168,40 +168,40 @@ void main() {
     n.stream = controller.stream;
 
     expect(data, null);
-    expect(n.data, null);
-    expect(n.isLoading, true);
-    expect(n.hasError, false);
-    expect(n.error, null);
-    expect(n.connectionState, ConnectionState.waiting);
-    expect(n, const AsyncSnapshot<int>.waiting());
+    expect(n.snapshot.data, null);
+    expect(n.snapshot.isLoading, true);
+    expect(n.snapshot.hasError, false);
+    expect(n.snapshot.error, null);
+    expect(n.snapshot.connectionState, ConnectionState.waiting);
+    expect(n.snapshot, const AsyncSnapshot<int>.waiting());
 
     controller.add(1);
     await Future<void>.delayed(const Duration(seconds: 1));
 
     expect(data, 1);
-    expect(n.data, 1);
-    expect(n.isReloading, true);
-    expect(n.hasError, false);
-    expect(n.error, null);
-    expect(n.connectionState, ConnectionState.active);
-    expect(n, AsyncSnapshot.withData(ConnectionState.active, data));
+    expect(n.snapshot.data, 1);
+    expect(n.snapshot.isReloading, true);
+    expect(n.snapshot.hasError, false);
+    expect(n.snapshot.error, null);
+    expect(n.snapshot.connectionState, ConnectionState.active);
+    expect(n.snapshot, AsyncSnapshot.withData(ConnectionState.active, data));
 
     controller.addError(e);
     await Future<void>.delayed(const Duration(seconds: 1));
 
     // as we cannot have both data and error, error replaces data.
-    expect(n.data, null);
-    expect(n.isLoading, true);
-    expect(n.hasError, true);
-    expect(n.error, error);
-    expect(n.connectionState, ConnectionState.active);
+    expect(n.snapshot.data, null);
+    expect(n.snapshot.isLoading, true);
+    expect(n.snapshot.hasError, true);
+    expect(n.snapshot.error, error);
+    expect(n.snapshot.connectionState, ConnectionState.active);
 
     final snapshotWithError = AsyncSnapshot<int>.withError(
       ConnectionState.active,
       error!,
       stackTrace!,
     );
-    expect(n, snapshotWithError);
+    expect(n.snapshot, snapshotWithError);
 
     controller.add(2);
     await controller.close();
@@ -209,12 +209,12 @@ void main() {
     await Future<void>.delayed(const Duration(seconds: 1));
 
     expect(data, 2);
-    expect(n.data, 2);
-    expect(n.isLoading, false);
-    expect(n.hasError, false);
-    expect(n.error, null);
-    expect(n.connectionState, ConnectionState.done);
-    expect(n, AsyncSnapshot.withData(ConnectionState.done, data));
+    expect(n.snapshot.data, 2);
+    expect(n.snapshot.isLoading, false);
+    expect(n.snapshot.hasError, false);
+    expect(n.snapshot.error, null);
+    expect(n.snapshot.connectionState, ConnectionState.done);
+    expect(n.snapshot, AsyncSnapshot.withData(ConnectionState.done, data));
   });
 
   test('AsyncNotifier.stream with new stream', () async {
@@ -231,29 +231,29 @@ void main() {
     // First stream.
     n.stream = Stream.periodic(const Duration(seconds: 1), (_) => throw error);
     expect(data, null);
-    expect(n.data, null);
-    expect(n.isLoading, true);
-    expect(n, const AsyncSnapshot<int>.waiting());
+    expect(n.snapshot.data, null);
+    expect(n.snapshot.isLoading, true);
+    expect(n.snapshot, const AsyncSnapshot<int>.waiting());
 
     // Second stream.
     n.stream = Stream.periodic(const Duration(seconds: 1), (i) => 1).take(1);
     expect(data, null);
-    expect(n.data, null);
-    expect(n.isLoading, true);
-    expect(n, const AsyncSnapshot<int>.waiting());
+    expect(n.snapshot.data, null);
+    expect(n.snapshot.isLoading, true);
+    expect(n.snapshot, const AsyncSnapshot<int>.waiting());
 
     // First stream has no time to throw. It's overriden by second stream.
     await Future<void>.delayed(const Duration(seconds: 1));
     expect(data, 1);
-    expect(n.data, 1);
-    expect(n.error, isNot(error));
-    expect(n.hasError, false);
-    expect(n.isLoading, true);
-    expect(n.isReloading, true);
-    expect(n, AsyncSnapshot.withData(ConnectionState.active, data));
+    expect(n.snapshot.data, 1);
+    expect(n.snapshot.error, isNot(error));
+    expect(n.snapshot.hasError, false);
+    expect(n.snapshot.isLoading, true);
+    expect(n.snapshot.isReloading, true);
+    expect(n.snapshot, AsyncSnapshot.withData(ConnectionState.active, data));
 
     await Future<void>.delayed(const Duration(seconds: 1));
-    expect(n, AsyncSnapshot.withData(ConnectionState.done, data));
+    expect(n.snapshot, AsyncSnapshot.withData(ConnectionState.done, data));
 
     // Error never called.
     expect(completer.isCompleted, false);
@@ -263,9 +263,10 @@ void main() {
     final n = AsyncNotifier<String>();
 
     String when() {
-      return n.when(
+      return n.snapshot.when(
         skipLoading: true,
-        data: (data) => '${n.isReloading ? 'reloading' : 'data'}: $data',
+        data: (data) =>
+            '${n.snapshot.isReloading ? 'reloading' : 'data'}: $data',
         error: (e, s) => 'error $e',
         loading: () => 'loading',
         none: () => 'none',
@@ -313,9 +314,10 @@ void main() {
     final state = AsyncNotifier<String?>();
 
     String when() {
-      return state.when(
+      return state.snapshot.when(
         skipLoading: true,
-        data: (data) => '${state.isReloading ? 'reloading' : 'data'}: $data',
+        data: (data) =>
+            '${state.snapshot.isReloading ? 'reloading' : 'data'}: $data',
         error: (e, s) => 'error $e',
         loading: () => 'loading',
         none: () => 'none',
@@ -359,10 +361,11 @@ void main() {
     final state = AsyncNotifier<void>();
 
     String when() {
-      return state.when(
+      return state.snapshot.when(
         skipLoading: true,
-        data: (_) => '${state.isReloading ? 'reloading' : 'data'}: ✅',
-        error: (e, s) => '${state.isReloading ? 'reloading' : 'error'}: $e',
+        data: (_) => '${state.snapshot.isReloading ? 'reloading' : 'data'}: ✅',
+        error: (e, s) =>
+            '${state.snapshot.isReloading ? 'reloading' : 'error'}: $e',
         loading: () => 'loading',
         none: () => 'none',
       );